put cleanup in another middleware
